api/requirements.txt
tensorflow==2.16.1
fastapi
uvicorn[standard]
pillow
numpy
python-multipart